rubocop 修正5

+++ db/migrate/20230721062956_add_price_to_items.rb
@@ -1,7 +1,5 @@
-# frozen_string_literal: true
-
 class AddPriceToItems < ActiveRecord::Migration[7.0]
   def change
-    add_column :items, :price, :decimal
+    add_column :items, :price, :integer
   end
 end
